+++ projects/signal-plus/src/lib/reactive-queries/sp-query.spec.ts
@@ -1,6 +1,7 @@
 import { signal } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-import { QueryClient, setGlobalQueryClient } from './query-client';
+import { QueryClient } from './query-client';
+import { setGlobalQueryClient } from './global-query-client';
 import { createDependentQuery, createQuery, spQuery } from './sp-query';
 
 describe('spQuery', () => {